Fix wrong display value for Selectors with static getters
The problem was that self.selection_idx was written from the clicked index and not updated through the synced value.
If the synced value would detect no change (e.g. because the getter was static -> non changing) the on_change would not be called to correctly replace self.selection_idx.